Refactor code: add cityList component
